api/requirements.txt
fastapi-cloud-cli==0.3.1
rignore==0.7.1
sentry-sdk==2.42.1
neo4j==6.0.2